Initial Integration of UI
