+++ Assets/Resources/Card/Strike.asset
@@ -10,20 +10,13 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: b327d123434b54ddb896c7e7201a3bd3, type: 3}
-  m_Name: Card 1
+  m_Name: Strike
   m_EditorClassIdentifier: 
   title: Strike
+  backgroundColor: {r: 0, g: 0, b: 0, a: 0}
   strategy: 2
   attribute: 0
   description: Something
   effectList:
-  - affectiveUnit: 1
-    diff: 2
-  - affectiveUnit: 1
-    diff: 2
-  - affectiveUnit: 1
-    diff: 2
-  - affectiveUnit: 1
-    diff: 2
-  - affectiveUnit: 1
-    diff: 2
+  - affectiveUnit: 2
+    damage: 3
