--- Assets/Scripts/Cards/CardFactory.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 76709873725af434bb82c568b2a9051a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
