+++ Assets/Scripts/UI/TestButton.cs
@@ -1,4 +1,7 @@
-﻿using UnityEngine.UI;
+﻿// Disable the entire script after reconstructure
+
+/*
+using UnityEngine.UI;
 using UnityEngine;
 using System.Collections.Generic;
 using Cards;
@@ -8,17 +11,19 @@ using Effects;
 public class TestButton : MonoBehaviour
 {
     CardFactory factory;
-
+    
     private void Start()
     {
         factory = FindObjectOfType<CardFactory>();
     }
-
-
+    
+    
     public void SpawnCard() { 
-
+    
         var effect = new Effects.Effect(Effects.UnitType.Enemy, 10);
         factory.MakeCard("Test", StrategyType.Berserker, AttributeType.Infernal, new List<Effect>() { effect });
     }
 
 }
+
+ */
